Use double precision in basta time

+++ cloudnetpy/instruments/basta.py
@@ -54,7 +54,7 @@ def basta2nc(
         if date is not None:
             basta.validate_date(date)
         basta.screen_data(keymap)
-        basta.add_time_and_range()
+        basta.add_time_and_range(time_dtype="f8")
         basta.add_site_geolocation()
         basta.add_zenith_angle()
         basta.add_radar_specific_variables()
